Match B2C course indexing for algolia
Replicates the logic from course-discovery that restricts which courses
are indexed for algolia search.
Some pieces of the logic look a little bit different as we use fields
coming from the API whereas they use fields in the discovery database.
Additionally, there is some logic in discovery that the website team
felt was duplicated that has been omitted here.

ENT-2992