Partial revert of 0b6d22e09f3580 -- Mac Bash complains about empty array
People on Mac were getting `edxapp_service_args[@]: unbound variable`
even though that was working on Linux; this is likely due to Apple
packaging an older version of Bash where `-u` treats empty arrays as unset.

This restores the use of an unquoted variable but adds a comment to warn
people away from quoting it.