Allow updating a specified subset of catalogs
Adds an optional kwarg `catalog_uuids` to the update_content_metadata
command which allows users to only update the ContentMetadata for the
catalogs that are associated with the catalog query of the provided
catalog.
Note that this does not change the update_full_content_metadata task to
only do that process for the subset of content associated with the
catalog queries of the enterprise catalogs that are given here. That is
being split out to try and make the refactoring cleaner.