Prefetch related objects in reindex-algolia task
Prefetches the associated catalog queries and their associated
enterprise catalogs to avoid the duplicate queries from ManyToMany and
ReverseForeignKey lookups from ContentMetadata --> CatalogQuery -->
EnterpriseCatalog.
This also drops the `enterprise_catalogs` property as that seemed to
circumvent some of Django's prefetching.
Additionally, this drops the `.values` call as that was clearing the
prefetched objects cache due to being a different query.