Update manual & auto catalog refresh functionality
* Updates the refresh_metadata endpoint for a catalog to perform all the
steps of `update_content_metadata`, `update_full_content_metadata`, and
`reindex_algolia` one after the other for the content associated with
the given catalog's CatalogQuery. This is done both to bring the
endpoint back in sync with the functionality of the aforementioned
scheduled jenkins jobs, as well as to give ECS and developers an easy
way to update a single CatalogQuery at a time (used in onboarding new
EnterpriseCustomers / EnterpriseCatalogs).
* Significant refactoring of the current tasks used by the jobs to
support both flows
* Minor quality and test fixes
* Have the update_content_metadata mgmt command block until the async result is ready,
or 30 minutes have passed, whichever comes first.

ENT-2450