fix: align catalog view tests with new default test POST format
First, this commit changes all instances of self.client.post() in unit
tests to NOT specify a content_type and NOT serialize data to JSON
wherever it previously did those things.  By default, rest_framework
will convert the input data to JSON, so there's no more need to do so
manually.

Second, since the default test POST format changed from `multipart` to
`json` it started to cause tests to fail, specifically when the data
object contained keys which pointed to string values that could be
deserialized into JSON and where the EnterpriseCatalogSerializer was
expecting a JSON object.  The tests had string literals hard-coded in
the POST data where there should have been python dicts, which caused
actual strings to get deserialized instead of dicts, so I converted
those test fields to from string literals to dicts.  Note that this does
actually change the behavior of the tests, by sending a POST requests
containing `application/json` rather than `multipart/form-data` bodies,
but I think this change brings the tests closer to the reality of what
the API caller does (which is to send JSON).