fix: reindex-algolia - include UUIDs from ALL related objects
Prior to this change we essentially looped over courses and appended
their UUIDs to related parent objects (programs and pathways), but now
we have to do this inverse: loop over higher level objects and absorb
UUIDs from all child objects.  This is a fundamental change to the
access pattern that would normally have no impact to the output since it
just changes the order in which UUIDs are accumulated; in the old world
where we accumulate all the UUIDs into a giant 16GB+ dictionary object,
the order they are added doesn't matter because they'd all get there by
the end.  However, now without being able to accumulate all UUIDs
in-memory, we have to essentially change the UUID access pattern to be
insensitive to constantly clearing the memory between batches.