fix: fix task-locking bug related to django-celery-results and protocol v2
ENT-10143: fixes TaskResult query inside task_recently_run(). We recently switched to
celery protocol 2, and django-celery-results gives preference to
`kwargsrepr` over `kwargs` (and similar for args/argsrepr) when persisting results. So we have
to json-serialize the repr() of the task args/kwargs in order for our lookup of
recent tasks with the same name and input. This is ok to do because none of the task
invocations in this code repo specify a custom args/kwargsrepr when submitting tasks, and
celery uses the repr() function as the default.