Fix  add task deleting old tmp indices (#1061)
Algolia gets more and more catalog tmp indices over time. This is due to Algolia's behavior to keep tmp indices when running replace_all_objects and encountering errors, which happens sometimes. This method is used by us when reindexing.

These tmp indices take up a large space in Algolia, and therefore should be regularly deleted.

This adds a management task that can delete these tmp indices. It filters for them by name to ensure only catalog tmp indices are deleted, and also only deletes tmp indices that are between 10 and 60 days old, to ensure that no indices that are new and in use are deleted.