chore: delete tmp indices task should depend on updated_at field (#1069)
This is related to a management task that is supposed to delete tmp indices from Algolia.

Testing deleting tmp indices on prod didn't work. An investigation showed that created_at for these tmp indices was set to the creation date of the original index they were copying, in 2020.
I changed the selection of date range for which indices to delete to use the updated_at field. As far as I can see, that corresponds to the Last Active display on the Algolia dashboard.
I introduced logging for each index including its metadata and added logging to track why it becomes marked for deletion or not. That way when testing on prod with dry run, we can exactly trace what is happening.
I added one more safeguard, just in case, that makes sure the index's entry field is 0 (telling us that the index is empty).