fix: remove tmp indices date check (#1072)
Minor fix to the date check of the remove-tmp-indices task.
Since the index date was < task min_days_ago, that didn't include the day specified.
For example, if you set min_days_ago to 0, that won't include today.
That makes it difficult to test index deletion on stage, so I'm fixing it here.