fix: normalized course price should never fallback to None
This fixes a bug where enterprise-access can-redeem returns 422 for an
ExecEd course without an advertised course run because it could not
tolerate a null price from enterprise-catalog. To solve that case, the
fix in this commit is to set the normalized price to the ExecEd
entitlement price instead of null. If the course is not an ExecEd
course, then fallback to 0.0 dollars instead of null.

ENT-10866